// arcadia.backoffice.fe/src/modules/dashboard/types.ts
export enum CurrentActivesColor {
  existing = '#32c5ff',
  new = '#e02020',
}

export enum CurrentActivesLabel {
  existing = 'Existing',
  new = 'New',
}
// ...
export interface ICurrentActivesChartItem {
  name: CurrentActivesLabel
  color: CurrentActivesColor
  value: number
}

export type ICurrentActivesChartData = ICurrentActivesChartItem[];